Restructure examples
framework_usage and testing are gone in favor of basic_structure and framework_features. The examples/README now describes what each example contains. And the root README points to the examples folder

[amend commit] fix rustfmt

[amend commit] fix some things to make the example actually runnable (lol), fix feature gates, use poise::builtins::register_globally
