+++ examples/feature_showcase/checks.rs
