+++ examples/feature_showcase/modal.rs
@@ -1,0 +1,17 @@
+use crate::{Data, Error};
+
+#[derive(Debug, poise::Modal)]
+#[allow(dead_code)] // fields only used for Debug print
+struct MyModal {
+    first_input: String,
+    second_input: Option<String>,
+}
+#[poise::command(slash_command)]
+pub async fn modal(ctx: poise::ApplicationContext<'_, Data, Error>) -> Result<(), Error> {
+    use poise::Modal as _;
+
+    let data = MyModal::execute(ctx).await?;
+    println!("Got data: {:?}", data);
+
+    Ok(())
+}
