+++ examples/feature_showcase/paginate.rs
@@ -1,0 +1,15 @@
+use crate::{Context, Error};
+
+#[poise::command(slash_command, prefix_command)]
+pub async fn paginate(ctx: Context<'_>) -> Result<(), Error> {
+    let pages = &[
+        "Content of first page",
+        "Content of second page",
+        "Content of third page",
+        "Content of fourth page",
+    ];
+
+    poise::samples::paginate(ctx, pages).await?;
+
+    Ok(())
+}
