--- examples/framework_usage/commands/mod.rs
@@ -1,6 +1,0 @@
-pub mod autocomplete;
-pub mod checks;
-pub mod context_menu;
-pub mod general;
-pub mod localization;
-pub mod subcommands;
